convert test position to int64, map index type of invalid position exceptions

+++ Nata.IO.Tests/ReaderTests.fs
@@ -17,7 +17,7 @@ type ReaderTests() =
           Data = ()
           Metadata = () }
 
-    abstract member Connect : unit -> List<Capability<'Data,'Metadata,int>>
+    abstract member Connect : unit -> List<Capability<'Data,'Metadata,int64>>
     
     member x.Connect(fn) =
         let stream = x.Connect()
